change_parameter added

+++ src/main.c
@@ -23,11 +23,11 @@ int main(int argc, char** argv) {
     uint64_t* fields = malloc(sizeof(uint64_t) + sizeof(char*));
     fields[0] = 1024;
     char* str = "billy";
-    fields[1] = str;
+    fields[1] = (uint64_t) str;
     add_tuple(file, fields, 1298);
     fields[0] = 123;
     str = "Bob";
-    fields[1] = str;
+    fields[1] = (uint64_t) str;
     add_tuple(file, fields, 3);
 
     print_tree_header_from_file(file);
